+++ rules/no-single-composition.js
@@ -8,13 +8,21 @@ const create = function (context) {
 
   return info.merge({
     CallExpression: function (node) {
-      if (node.arguments.length > 1 || _.get([0, 'type'], node.arguments) === 'SpreadElement') {
+      const method = info.helpers.isComposeMethod(node);
+      if (!method || node.arguments.length > 1 || _.get([0, 'type'], node.arguments) === 'SpreadElement') {
         return;
       }
-      const method = info.helpers.isComposeMethod(node);
-      if (method) {
-        context.report(node, `\`${method.name}\` should have at least two arguments`);
+
+      // _.flow accepts array of Functions, check that we have more than one Function
+      const isArrayExpression = _.get([0, 'type'], node.arguments) === 'ArrayExpression';
+      if (isArrayExpression && _.get([0, 'elements', 'length'], node.arguments) > 1) {
+        return;
       }
+
+      context.report({
+        node,
+        message: `\`${method.name}\` should have at least two ${isArrayExpression ? 'functions' : 'arguments'}`
+      });
     }
   });
 };